フォロー機能の追加および各不具合の修正

+++ src/models/Invitation.js
@@ -15,6 +15,8 @@ class Invitation {
     const invitation = db.collection('invitation').doc(email);
     const document = await invitation.get();
 
+    if (!document.exists) return;
+
     User.rewardForInviteUser(document.data().uid);
     invitation.delete();
   };
